deduplicate psiexpression methods
